0.3.1 snapshot

+++ FinalProject/src/main/java/team/last/project/repository/ReserveRepository.java
@@ -4,10 +4,9 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import team.last.project.entity.OptPrice;
 import team.last.project.entity.Reserve;
 public interface ReserveRepository extends JpaRepository<Reserve, Integer>{
 
-	List<Reserve> findALLByMemberId(Integer id);
+	List<Reserve> findALLByMemberId(Long id);
 
 }
